polymarket-trading-bot/requirements.txt
# Polymarket客户端
py-clob-client>=0.20.0

# Web3交互
web3>=6.0.0
eth-account>=0.10.0

# 异步框架
aiohttp>=3.9.0
aiofiles>=23.0.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 数据库
SQLAlchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# 配置管理
python-dotenv>=1.0.0
PyYAML>=6.0.0
pydantic>=2.0.0

# 日志
loguru>=0.7.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# 监控和告警(可选)
sentry-sdk>=1.30.0

# 实用工具
requests>=2.31.0
python-dateutil>=2.8.0
pytz>=2023.3
